pyproject.toml
[project]
name = "cursor-practices"
version = "0.1.0"
description = "FastAPI application for random number generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]
